Move internal headers into snmp_internal
