+++ src/snmp_scalar.cpp
@@ -36,7 +36,7 @@
 
 #include <string>
 
-#include "snmp_includes.h"
+#include "snmp_internal/snmp_includes.h"
 #include "snmp_scalar.h"
 
 namespace SNMP
